refactor(ui): move QoS axis time formatting into datetime helper
Keep chart labels on the shared timezone path instead of exporting a helper from the untyped QoS panel.
