+++ src/ui/src/lib/datetime.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { withTimeZone } from './datetime'
+import { formatAxisTime, withTimeZone } from './datetime'
 
 describe('withTimeZone', () => {
   it('strips pre-existing option timezone when local is requested', () => {
@@ -13,5 +13,31 @@ describe('withTimeZone', () => {
       hour: '2-digit',
       timeZone: 'America/New_York',
     })
+  })
+})
+
+const axisOptions = {
+  hour: 'numeric',
+  minute: 'numeric',
+  second: 'numeric',
+} satisfies Intl.DateTimeFormatOptions
+
+describe('formatAxisTime', () => {
+  it('uses the browser/runtime timezone when local is selected', () => {
+    const date = new Date('2026-08-17T11:30:04Z')
+    const unixSec = date.getTime() / 1000
+
+    expect(formatAxisTime(unixSec, 'local', 'fr-FR')).toBe(
+      new Intl.DateTimeFormat('fr-FR', axisOptions).format(date),
+    )
+  })
+
+  it('uses an explicit timezone when one is selected', () => {
+    const date = new Date('2026-08-17T11:30:04Z')
+    const unixSec = date.getTime() / 1000
+
+    expect(formatAxisTime(unixSec, 'UTC', 'fr-FR')).toBe(
+      new Intl.DateTimeFormat('fr-FR', { ...axisOptions, timeZone: 'UTC' }).format(date),
+    )
   })
 })
